Extract TestPairSearch from agent_test.go
to separate file.
